requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alpha_vantage==3.0.0
anyio==4.4.0
appdirs==1.4.4
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
frozendict==2.4.4
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.2
idna==3.8
ldap3==2.9
lxml==5.2.2
multidict==6.0.5
multitasking==0.0.11
numpy==2.1.1
pandas==2.2.2
pandas-datareader==0.10.0
peewee==3.17.6
platformdirs==4.2.2
pyasn1==0.6.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-telegram-bot==21.2
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.30
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
yarl==1.9.11
